Cinema class removed and relation added
Cinema class only made sense just if we wanted to manage different cinemas from the same company. But our objective is to manage a single cinema, so I don't think it is necessary. If in the future we need that class we will need to implement it using singleton pattern.

Relation between Screen and Seat has been added.